requirements.txt
python-pptx==0.6.21
pandas==2.1.4
matplotlib==3.8.2
numpy==1.26.2
seaborn==0.13.0
openpyxl==3.1.2
